Dismiss notifications remotely (v2)
Just send a push with the GCM variable cancelNotification set (can be set to anything) and the notification will be dismissed. Default functionality uses the standard NOTIFICATION_ID=237.

The code addition also works with the the code from TheNewMR (thenewmr@728e975) which allows for stacked notifications depending on different notification ID's. To dismiss a Notification ID just make sure to set the GCM variable cancelNotification and to set notId to the notification ID you want to dismiss.